Cleanup & Code style
