--- packages/govern-console/src/lib/query-types.js
@@ -1,2 +1,0 @@
-export const BUY_ORDER = Symbol('BUY_ORDER')
-export const SELL_ORDER = Symbol('SELL_ORDER')
